refactor(newspaper): docx is import source, only ship FB2
Editor feedback on the previous docx round (PR #72): the .docx is
the source format, NOT something readers should download. Now:

- DOCX upload (admin side) converts client-side to FB2 and only the
  .fb2 lands in the content repo. The .docx never enters the bucket.
- Direct FB2 upload (admin side) is a sibling option for editors who
  already have a .fb2 from another tool.
- Public-website integration becomes `newspaper-fb2`: just copies
  *.fb2 from src/content/newspaper/<slug>/assets to dist, mirroring
  newspaper-pdfs.
- NewspaperCard drops the DOCX download link. Keeps PDF + (optional)
  FB2 link, gated on whether a .fb2 exists in the source assets.
- Drop mammoth from public-website deps and the parse5-based
  html-to-fb2 transform — both moved to admin-website where the
  conversion now happens client-side via DOMParser.
