+++ package.json
@@ -38,7 +38,6 @@
     "husky": "^9.1.7",
     "lighthouse": "^13.0.3",
     "lint-staged": "^16.2.7",
-    "mammoth": "^1.12.0",
     "playwright-lighthouse": "^4.0.0",
     "typescript": "^5.9.3"
   },
